[security][CVE-2022-27809] Builtins should always take int64_t, not int
Builtins are allowed to take 32 or 64-bit integers
(int/int32_t versus int64_t). It's not clear why we allow this, as
integers from Hack are always 64-bit. A builtin taking an int will
just have the upper 32-bits truncated. Depending on the value, this
might change a negative number to a positive one, defeating various
input validity checks.
Remove support for 32-bit integers. Change all of the builtins which
took or returned an int to use int64_t instead. This required a few
fixups here and there, mainly around format strings. If the builtin
uses 32-bits ints internally, I left them. At least the truncation is
now obvious.

Test Plan:
New unit test. Beforehand this large negative input would
be turned into a large positive number, bypassing the "input cannot be
negative" check. Now that it takes int64_t, it is correctly recognized
as a negative number.

---
Initializes m_arr in ArrayInitBase

ArrayInitBase's protected constructor does not initialize
m_arr (presumably because a derived class is meant to do so). However
if a derived class' constructor throws, we'll run the ArrayInitBase
destructor, which will try to release m_arr. If m_arr hasn't been
initialized yet, it can contain garbage and we'll try to release
random memory. The protected constructor should initialize m_arr to
nullptr.

Test Plan:
New unit test which previously failed with ASAN builds

+++ hphp/runtime/ext/bz2/ext_bz2.cpp
@@ -79,12 +79,12 @@ bool HHVM_FUNCTION(bzclose, const Resource& bz) {
   return HHVM_FN(fclose)(bz);
 }
 
-Variant HHVM_FUNCTION(bzread, const Resource& bz, int length /* = 1024 */) {
+Variant HHVM_FUNCTION(bzread, const Resource& bz, int64_t length /* = 1024 */) {
   return HHVM_FN(fread)(bz, length);
 }
 
 Variant HHVM_FUNCTION(bzwrite, const Resource& bz, const String& data,
-                               int length /* = 0 */) {
+                               int64_t length /* = 0 */) {
   return HHVM_FN(fwrite)(bz, data, length);
 }
 
@@ -176,8 +176,8 @@ Variant HHVM_FUNCTION(bzerrno, const Resource& bz) {
   return f->errnu();
 }
 
-Variant HHVM_FUNCTION(bzcompress, const String& source, int blocksize /* = 4 */,
-                                  int workfactor /* = 0 */) {
+Variant HHVM_FUNCTION(bzcompress, const String& source, int64_t blocksize /* = 4 */,
+                                  int64_t workfactor /* = 0 */) {
   unsigned int source_len, dest_len;
 
   source_len = source.length();
@@ -197,7 +197,7 @@ Variant HHVM_FUNCTION(bzcompress, const String& source, int blocksize /* = 4 */,
   }
 }
 
-Variant HHVM_FUNCTION(bzdecompress, const String& source, int small /* = 0 */) {
+Variant HHVM_FUNCTION(bzdecompress, const String& source, int64_t small /* = 0 */) {
   int source_len = source.length();
   int error;
   uint64_t size = 0;
